Address warnings/errors from Rubocop
It appears that Rubocop didn't add support for Ruby 2.5 until v0.52.0.
We were locked to v0.51.0. As such, when you set `.rubocop.yml` to
a `TargetRubyVersion: 2.5`, there were errors/warnings emitted into
stderr about it.

This change updates to the latest version of Rubocop, v0.58.2, which
supports the latest Ruby versions and is likely more in-line with what
people expect.